After rebasing, I added changes that I had made after the commit I reset back to. Now, everything should be as it was for the original push
