+++ components/ApplicationCard.tsx
@@ -5,7 +5,7 @@ export default function ApplicationCard() {
     <div className="flex w-55 flex-col rounded-2xl border-2 border-[#1E4240] bg-white">
       <div className="flex flex-col gap-y-30">
         <div className="flex flex-row justify-end pt-3 pr-3">
-          <Button variant="applicationCard">...</Button>
+          <Button variant="tertiary">...</Button>
         </div>
 
         <div className="flex flex-col pb-3 pl-3">
