+++ components/MainDashboard.tsx
@@ -1,6 +1,5 @@
 'use client';
 
-import { useApplicationContext } from '@/contexts/ApplicationContext';
 import { ButtonLink } from './Button';
 import LogoutButton from './LogoutButton';
 import TabsDemo from './Tabs';
@@ -24,7 +23,7 @@ export default function MainDashboard() {
           </div>
 
           <div>
-            <ButtonLink href="/app/1234567890" variant="applicationMainPage">
+            <ButtonLink href="/app/1234567890" variant="tertiary">
               New Application
             </ButtonLink>
           </div>
